Fix NextAuth route exports and build compatibility
