+++ apps/web/app/api/lp/data/route.ts
@@ -1,4 +1,5 @@
 import { getSession } from "@/lib/auth";
+import { type Role } from "@/lib/auth-helpers";
 import { computeMetrics, loadPartnerInvestmentRecords } from "@/lib/lp-server";
 
 export const runtime = "nodejs";
@@ -7,12 +8,13 @@ export const dynamic = "force-dynamic";
 export async function GET() {
   try {
     const session = await getSession();
-    const email = session?.user?.email;
-    if (!session || !email) {
+    const user = session?.user;
+    const email = user?.email;
+    if (!session || !user || !email) {
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
 
-    const role = (session.user.role as "admin" | "lp" | "partner" | undefined) ?? "lp";
+    const role = (user.role as Role | undefined) ?? "lp";
     const { records } = await loadPartnerInvestmentRecords(email, role);
     const metrics = computeMetrics(records);
